+++ src/layouts/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import _ from 'lodash'
-import { Spin } from 'antd'
 import { getPersistor } from '@/utils/configureStore'
 import { PersistGate } from 'redux-persist/integration/react'
 import { connect } from 'dva'
@@ -8,24 +7,21 @@ import DocumentTitle from 'react-document-title'
 import { getLocaleContext } from 'kuu-tools'
 import BlankLayout from './BlankLayout'
 import BasicLayout from './BasicLayout'
+import Loading from '@/components/sys/loading'
 import config from '@/config'
 
 const IndexLayout = props => {
   const isSimple = config.simplePages.indexOf(props.location.pathname) >= 0
-  const LayoutContainer = isSimple ? BlankLayout : BasicLayout
   const LocaleContext = getLocaleContext()
-  const loadingStyle = {
-    width: '100vw',
-    height: '100vh',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center'
+  let LayoutContainer = isSimple ? BlankLayout : BasicLayout
+  if (props.loading) {
+    LayoutContainer = Loading
   }
   return (
     <>
       <PersistGate
         persistor={getPersistor()}
-        loading={<Spin style={loadingStyle} size='large' />}
+        loading={null}
       >
         <DocumentTitle title={config.htmlTitle}>
           <LocaleContext.Provider value={props.localeMessages}>
@@ -40,5 +36,8 @@ const IndexLayout = props => {
 }
 
 export default connect(state => {
-  return { localeMessages: _.get(state, 'i18n.localeMessages') }
+  return {
+    localeMessages: _.get(state, 'i18n.localeMessages'),
+    loading: _.get(state, 'loading.effects["layout/init"]') || _.get(state, 'loading.models.user') || _.get(state, 'loading.models.i18n')
+  }
 })(IndexLayout)
